# backend/src/main/java/net/mbope/taskmanager/user/internal/presentation/PasswordController.java
package net.mbope.taskmanager.user.internal.presentation;

import io.swagger.v3.oas.annotations.responses.ApiResponse;

import io.swagger.v3.oas.annotations.Operation;

import net.mbope.taskmanager.user.UserCredentials;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/users/me")
class PasswordController {
    private final UserCredentials credentials;

    PasswordController(UserCredentials credentials) {
        this.credentials = credentials;
    }

    @PutMapping(path = "/password", consumes = MediaType.APPLICATION_JSON_VALUE)
    @Operation(operationId = "changePassword")
    @ApiResponse(responseCode = "204")
    ResponseEntity<Void> changePassword(@RequestBody ChangePasswordRequest request) {
        credentials.changePassword(request.currentPassword(), request.newPassword());
        return ResponseEntity.noContent().build();
    }
}