# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Manager API",
    "version": "0.1.0-draft",
    "description": "Draft v1 API contract. Account HTTP endpoints and JWT authentication are implemented. Task domain and application use cases are implemented and tested through ports; task persistence and runtime wiring are implemented with H2 integration coverage. Personal and administrative task HTTP endpoints are implemented. PostgreSQL migration, constraint, filtering, and audit rollback tests pass; the full clean build passes 207 tests. Runtime OpenAPI 3.1 and Swagger UI are implemented, with independent contract comparison and profile-controlled documentation access. Based on ADR-0001 through ADR-0007. See docs/api/README.md for design choices, lifecycle, error identifiers, and implementation requirements. All request objects reject unknown properties. All application-controlled API errors use application/problem+json. Authentication is bearer-header-only; no authentication cookies. Deployed environments require HTTPS."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin base path; configure host per environment."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "summary": "Register an enabled USER account",
        "description": "Reject client-supplied roles, enabled, or IDs. Does not log in or issue a token.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created with USER role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        }
      }
    },
...
    "/users/me/password": {
      "put": {
        "tags": [
          "Accounts"
        ],
        "operationId": "changePassword",
        "summary": "Change the caller's password",
        "description": "Verify currentPassword against the stored hash. A mismatch returns 400 validation with field currentPassword and code incorrect_password. Existing JWTs retain their authority until expiry; this operation does not revoke them.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        }
      }
    },
...
    "/tasks/{taskId}": {
      "parameters": [
        {
          "name": "taskId",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Id"
          }
        }
      ],
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "getTask",
        "summary": "Get a task assigned to the caller",
        "responses": {
          "200": {
            "description": "Task details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "description": "Only the assignee may access a task through this personal route, including when the caller is ADMIN. Return identical 404 responses for missing tasks and tasks assigned to another account. Administrators use /admin/tasks/{taskId} for other tasks."
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "operationId": "updateTask",
        "summary": "Partially update a task assigned to the caller",
        "description": "Only the assignee may access a task through this personal route, including when the caller is ADMIN. Return identical 404 responses for missing tasks and tasks assigned to another account. Administrators use /admin/tasks/{taskId} for other tasks. Any status transition is allowed. Last write wins. Assignee and creator are immutable; reassignment is outside v1.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "operationId": "deleteTask",
        "summary": "Permanently delete a task assigned to the caller",
        "description": "Only the assignee may access a task through this personal route, including when the caller is ADMIN. Return identical 404 responses for missing tasks and tasks assigned to another account. Administrators use /admin/tasks/{taskId} for other tasks. Repeated deletion returns 404.",
        "responses": {
          "204": {
            "description": "Task deleted; no response body."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        }
      }
    },
...
    "/admin/tasks": {
      "post": {
        "tags": [
          "Administration"
        ],
        "operationId": "createAssignedTask",
        "summary": "Create a task assigned to a USER or ADMIN (ADMIN required)",
        "description": "Check ADMIN in token claims before looking up assigneeId. The assignee must be an existing enabled account, regardless of USER or ADMIN role. Return 404 for a missing account and 409 for a disabled account. Record the caller as createdById. The assignee manages the task through /tasks; any ADMIN can manage it through /admin/tasks. Assignment is synchronous; events and notifications remain deferred. Reassignment after creation is outside v1.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "headers": {
              "Location": {
                "required": true,
                "description": "Relative URL /api/v1/admin/tasks/{taskId}; accessible to the creating administrator even when assigned to someone else.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/AssigneeDisabled"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "x-required-roles": [
          "ADMIN"
        ]
      },
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listAllTasks",
        "summary": "List all tasks (ADMIN required)",
        "description": "Check ADMIN in token claims before querying. Includes all tasks regardless of creator or assignee, including self-created tasks and tasks assigned to disabled accounts. Optional assigneeId, status, and q filters combine with AND; calculate totals after filtering. A well-formed assigneeId with no matching tasks returns an empty page. Sort by createdAt descending, then id ascending.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Zero-based page index. Beyond the last page returns empty items."
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Maximum items per page."
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TaskStatus"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "pattern": "\\S"
            },
            "description": "Case-insensitive literal substring match on title; trim surrounding whitespace. SQL wildcards have no special meaning."
          },
          {
            "name": "assigneeId",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Id"
            },
            "description": "Filter tasks by assigned account UUID."
          }
        ],
        "responses": {
          "200": {
            "description": "All tasks matching filters, including an empty page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "x-required-roles": [
          "ADMIN"
        ]
      }
    },
    "/admin/tasks/{taskId}": {
      "parameters": [
        {
          "name": "taskId",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/Id"
          }
        }
      ],
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "getAnyTask",
        "summary": "Get any task (ADMIN required)",
        "responses": {
          "200": {
            "description": "Task details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "description": "Check ADMIN in token claims before task lookup. Any administrator may access the task regardless of creator, assignee, or assignee enabled status. Return 404 if the task does not exist.",
        "x-required-roles": [
          "ADMIN"
        ]
      },
      "patch": {
        "tags": [
          "Administration"
        ],
        "operationId": "updateAnyTask",
        "summary": "Partially update any task (ADMIN required)",
        "description": "Check ADMIN in token claims before task lookup. Any administrator may access the task regardless of creator, assignee, or assignee enabled status. Return 404 if the task does not exist. Any status transition is allowed. Last write wins. Assignee and creator remain immutable; reassignment is outside v1. Record the administrator's edit in a synchronous audit record.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "x-required-roles": [
          "ADMIN"
        ]
      },
      "delete": {
        "tags": [
          "Administration"
        ],
        "operationId": "deleteAnyTask",
        "summary": "Permanently delete any task (ADMIN required)",
        "description": "Check ADMIN in token claims before task lookup. Any administrator may access the task regardless of creator, assignee, or assignee enabled status. Return 404 if the task does not exist. Repeated deletion returns 404. Record the administrator's deletion in a synchronous audit record that survives task deletion.",
        "responses": {
          "204": {
            "description": "Task deleted; no response body."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "default": {
            "$ref": "#/components/responses/OtherError"
          }
        },
        "x-required-roles": [
          "ADMIN"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT with 900-second lifetime. Validate signature, algorithm, issuer, audience, and expiration. Use roles from token; do not recheck enabled status to revoke existing tokens."
      }
    },
    "schemas": {
      "Id": {
        "type": "string",
        "format": "uuid"
      },
      "NewPassword": {
        "type": "string",
        "format": "password",
        "writeOnly": true,
        "minLength": 15,
        "maxLength": 72,
        "description": "15 or more Unicode code points and at most 72 UTF-8 bytes. Enforce the byte limit server-side; maxLength alone is insufficient. Preserve spaces and Unicode exactly; do not trim or normalize."
      },
      "CurrentPassword": {
        "type": "string",
        "format": "password",
        "writeOnly": true,
        "minLength": 1,
        "maxLength": 72,
        "description": "Preserve input exactly. Reject values exceeding 72 UTF-8 bytes."
      },
      "Email": {
        "type": "string",
        "format": "email",
        "maxLength": 254,
        "description": "Trim surrounding whitespace and lowercase before storage and lookup; uniqueness is case-insensitive. Account email cannot be changed in v1."
      },
...
      "Account": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "roles",
          "enabled",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "email": {
            "$ref": "#/components/schemas/Email"
          },
          "roles": {
            "$ref": "#/components/schemas/Roles"
          },
          "enabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "$ref": "#/components/schemas/Email"
          },
          "password": {
            "$ref": "#/components/schemas/NewPassword"
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "$ref": "#/components/schemas/CurrentPassword"
          },
          "newPassword": {
            "$ref": "#/components/schemas/NewPassword"
          }
        }
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "TODO",
          "IN_PROGRESS",
          "DONE"
        ],
        "description": "Any transition between these states, including reopening DONE, is allowed. Setting the same state is allowed."
      },
...
      "Task": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "description",
          "status",
          "dueDate",
          "createdAt",
          "updatedAt",
          "assigneeId",
          "createdById"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Id"
          },
          "title": {
            "$ref": "#/components/schemas/TaskTitle"
          },
          "description": {
            "$ref": "#/components/schemas/TaskDescription"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "dueDate": {
            "$ref": "#/components/schemas/DueDate"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assigneeId": {
            "$ref": "#/components/schemas/Id",
            "description": "Account responsible for the task. Set to the caller for ordinary creation or the selected recipient for admin creation. Immutable in v1."
          },
          "createdById": {
            "$ref": "#/components/schemas/Id",
            "description": "Authenticated account that created the task. Immutable and never client-supplied."
          }
        }
      },
...
      "UpdateTaskRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [],
        "properties": {
          "title": {
            "$ref": "#/components/schemas/TaskTitle"
          },
          "description": {
            "$ref": "#/components/schemas/TaskDescription"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "dueDate": {
            "$ref": "#/components/schemas/DueDate"
          }
        },
        "minProperties": 1,
        "description": "Partial update. Omitted fields remain unchanged; null clears description or dueDate. title and status cannot be null. assigneeId, createdById, owner IDs, and resource IDs cannot be supplied. Reassignment is outside v1."
      },
...
      "TaskPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page",
          "size",
          "totalElements",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "totalElements": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "AdminCreateTaskRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "assigneeId"
        ],
        "properties": {
          "title": {
            "$ref": "#/components/schemas/TaskTitle"
          },
          "description": {
            "$ref": "#/components/schemas/TaskDescription",
            "default": null
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "TODO"
          },
          "dueDate": {
            "$ref": "#/components/schemas/DueDate",
            "default": null
          },
          "assigneeId": {
            "$ref": "#/components/schemas/Id",
            "description": "Existing enabled account receiving this task. May be a USER or ADMIN, including the caller."
          }
        },
        "description": "Create and assign a new task. The authenticated ADMIN is recorded as creator; the assignee is selected explicitly."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid JSON, unknown fields, malformed parameters, or constraint violations. Includes field errors.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblem"
            },
            "example": {
              "type": "urn:task-manager:problem:validation",
              "title": "Bad Request",
              "status": 400,
              "detail": "Invalid JSON, unknown fields, malformed parameters, or constraint violations. Includes field errors.",
              "instance": "/api/v1/tasks",
              "errors": [
                {
                  "field": "$",
                  "code": "invalid",
                  "message": "Request is invalid."
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:unauthorized",
              "title": "Unauthorized",
              "status": 401,
              "detail": "Missing, invalid, or expired bearer token.",
              "instance": "/api/v1/tasks"
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "required": true,
            "description": "Bearer challenge; use Bearer for missing credentials and Bearer error=\"invalid_token\" for invalid or expired tokens.",
            "schema": {
              "type": "string"
            },
            "example": "Bearer"
          }
        }
      },
...
      "Forbidden": {
        "description": "The token does not contain ADMIN.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:forbidden",
              "title": "Forbidden",
              "status": 403,
              "detail": "The token does not contain ADMIN.",
              "instance": "/api/v1/tasks"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found. Missing and inaccessible tasks use identical responses.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:not-found",
              "title": "Not Found",
              "status": 404,
              "detail": "Resource not found.",
              "instance": "/api/v1/tasks"
            }
          }
        }
      },
      "Conflict": {
        "description": "An account with this normalized email already exists.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:email-conflict",
              "title": "Conflict",
              "status": 409,
              "detail": "An account with this normalized email already exists.",
              "instance": "/api/v1/tasks"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Request content type must be application/json.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:unsupported-media-type",
              "title": "Unsupported Media Type",
              "status": 415,
              "detail": "Request content type must be application/json.",
              "instance": "/api/v1/tasks"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error; no internal details are disclosed.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:internal-error",
              "title": "Internal Server Error",
              "status": 500,
              "detail": "Unexpected server error; no internal details are disclosed.",
              "instance": "/api/v1/tasks"
            }
          }
        }
      },
      "OtherError": {
        "description": "Other application-controlled error. Use Problem Details with status matching the HTTP response. Unsupported methods return 405 with urn:task-manager:problem:method-not-allowed and an Allow header. Infrastructure errors may be outside this contract.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AssigneeDisabled": {
        "description": "The selected account is disabled and cannot receive a new assignment.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "urn:task-manager:problem:assignee-disabled",
              "title": "Conflict",
              "status": 409,
              "detail": "The selected account is disabled.",
              "instance": "/api/v1/admin/tasks"
            }
          }
        }
      }
    }
  }
}